+++ rrule-debugger/src/parser_rrule.rs
@@ -18,7 +18,7 @@ pub fn parse_rrule_from_string(rrule: &str) {
         Ok(rrule) => {
             println!("RRule data: {:#?}", rrule);
             let result = rrule.all(50);
-            crate::print_all_datetimes(&result.list);
+            crate::print_all_datetimes(&result.dates);
             if result.limited {
                 println!("RRule was limited");
             }